% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{active_access}
\alias{active_access}
\title{Utility function for R6 classes}
\usage{
active_access(priv, warn.msg = "Field cannot be modified")
}
\arguments{
\item{priv}{private object to return}

\item{warn.msg}{warning message if user tries to modify this field}
}
\description{
Utility function for R6 classes
}
