% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapBlock.R
\name{calcBlocks}
\alias{calcBlocks}
\title{Helper function to compute blocks of markers}
\usage{
calcBlocks(pos, window = NULL, n = NULL, exclude.size = 10)
}
\arguments{
\item{pos}{data.frame containing at least columns "CHROM" and "POS"
containing chromosome and position location}

\item{window}{numeric, window size. Markers are grouped in windows of `window` size.}

\item{n}{numeric, number of markers to be grouped together.}

\item{exclude.size}{numeric, if a block contains less than `exclude.size` it will be omitted.
Defaults to 10.}
}
\value{
a split data.frame. A column named `index` is added, which contains the
original row index (and is useful to subset the `phase` data.frame)
}
\description{
Helper function to compute blocks of markers
}
\examples{
splitPos <- calcBlocks(exPos,n = 20)

myBlock <- hapBlock$new(exPhase[splitPos[[1]]$index,],exInds)
}
