% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exInds}
\alias{exInds}
\title{Individual labels}
\format{
## `exInds`
A character vector with individual names.
}
\usage{
exInds
}
\description{
Example individual labels. Anonimized labels related to the `exPhase` dataset.
The order of the labels relates to the order of the individuals in the `exPhase` data.
}
\keyword{datasets}
