% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exPhase}
\alias{exPhase}
\title{Phase data}
\format{
## `exPhase`
A data frame with 1229 rows and 2458 columns. Each column is one chromosome
of a single individual, so there are ploidy*individuals columns.
}
\usage{
exPhase
}
\description{
Example phased SNP data, in binary form, with individuals in columns
and markers in rows. The dataset is a subset of a genomic SNP-phased data
from the Center for Genomic Resources Netherlands (CGN).
}
\keyword{datasets}
