% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exPos}
\alias{exPos}
\title{Position of markers}
\format{
## `exPos`
A data.frame with 4 columns and 1229 rows, one for each marker in the `exPhase` dataset.
}
\usage{
exPos
}
\description{
Example position information related to the `exPhase` dataset. The data.frame contains
four columns, "CHROM" (chromosome name), "POS" (genomic position, in bp), "REF" (reference
allele) and "ALT" (alternative allele)
}
\keyword{datasets}
