% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapBlock.R
\name{hapBlock}
\alias{hapBlock}
\title{R6 Class Representing haploBlock}
\description{
A haploBlock is a set of phased markers across several individuals
that can be analyzed together.
}
\details{
Several methods are implemented within the `hapBlock` object, specifically
used to
 1) Group similar variants into 'haplogroups'
 2) Obtain statistics about variants and haplogroups
 3) Obtain allele information in various formats
 (Ref/Alt, variant codes, haplotype codes...)
 4) Visualize variant grouping
}
\examples{

## ------------------------------------------------
## Method `hapBlock$new`
## ------------------------------------------------


hapBlock$new(exPhase,exInds, ploidy = 2)


## ------------------------------------------------
## Method `hapBlock$haplogroup`
## ------------------------------------------------


myBlock <- hapBlock$new(exPhase,exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)


## ------------------------------------------------
## Method `hapBlock$haplostats`
## ------------------------------------------------

myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)
stats <- myBlock$haplostats()

## ------------------------------------------------
## Method `hapBlock$plotGrouping`
## ------------------------------------------------

myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)
myBlock$plotGrouping("My plot")

## ------------------------------------------------
## Method `hapBlock$alleles`
## ------------------------------------------------

myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$alleles("var")

myBlock$haplogroup(error = 0.02)
myBlock$alleles("hap")

## ------------------------------------------------
## Method `hapBlock$mismatch`
## ------------------------------------------------

myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
expected <- myBlock$mismatch(error = c(0.01))
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{inds}}{individuals in the population (each repeated `ploidy` times)}

\item{\code{phase}}{phase dataset provided}

\item{\code{codes}}{translation table between different allele encodings (binary, variant, haplogroup...)}

\item{\code{D}}{distance matrix with mismatches between observed variants}

\item{\code{pos}}{data.frame with position information, if given (NULL otherwise)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hapBlock-new}{\code{hapBlock$new()}}
\item \href{#method-hapBlock-print}{\code{hapBlock$print()}}
\item \href{#method-hapBlock-haplogroup}{\code{hapBlock$haplogroup()}}
\item \href{#method-hapBlock-haplostats}{\code{hapBlock$haplostats()}}
\item \href{#method-hapBlock-plotGrouping}{\code{hapBlock$plotGrouping()}}
\item \href{#method-hapBlock-alleles}{\code{hapBlock$alleles()}}
\item \href{#method-hapBlock-buildSequences}{\code{hapBlock$buildSequences()}}
\item \href{#method-hapBlock-mismatch}{\code{hapBlock$mismatch()}}
\item \href{#method-hapBlock-clone}{\code{hapBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-new}{}}}
\subsection{Method \code{new()}}{
Create a new haploBlock object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$new(phase, inds, ploidy = 2, pos = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phase}}{A matrix where each row is a marker and each column is
a single chromosome of one individual (a single, phased chromosome)}

\item{\code{inds}}{A vector of individual names (not column names). It is assumed
that individual order and column order are the same.}

\item{\code{ploidy}}{Numeric, the ploidy of the organism. The number of inds * ploidy
should be equal to the number of columns in phase.}

\item{\code{pos}}{A data.frame with marker position information. More specifically
with columns "chrom", "pos", "ref", "alt", containing the chromosome location,
base-pair position, reference allele and alternative allele of each marker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`hapBlock` object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
hapBlock$new(exPhase,exInds, ploidy = 2)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-print"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-print}{}}}
\subsection{Method \code{print()}}{
Print method for a hapBlock object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-haplogroup"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-haplogroup}{}}}
\subsection{Method \code{haplogroup()}}{
Grouping phased variants into similarity groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$haplogroup(error = 0.01, mismatch = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error}}{Numeric, expected genotyping errors in the phase data. By default 0.01}

\item{\code{mismatch}}{Number of expected mismatches. If provided it will ignore the error parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Only a single 'grouping' can be stored at a time. If several groupings are performed
one after the other, only the last will be stored within the object.
}

\subsection{Returns}{
Does not return a value, but other methods of hapBlock will be able to return results
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
myBlock <- hapBlock$new(exPhase,exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-haplostats"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-haplostats}{}}}
\subsection{Method \code{haplostats()}}{
Statistics of haplotype grouping. Can only be used after using the method `haplogroup()`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$haplostats()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame with summary statistics of the haplogroups:
- group: haplogroup code, numbered from most to least frequent
- meanD: mean distance (mismatches) between variants within the group
- nvars: number of distinct variants within the group
- groupFreq: number of times the group is observed in the total population
- consensus: consensus variant, where each position is the most frequent. If less than 3/4 of the
observed variants have the same allele in a position, an N is returned instead.
- freqVar: the most frequent variant within the haplogroup
- freq: the frequency, within the haplogroup, of the most frequent variant
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)
stats <- myBlock$haplostats()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-plotGrouping"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-plotGrouping}{}}}
\subsection{Method \code{plotGrouping()}}{
Visualization of the obtained haplotype grouping. Each vertex (dot) is a variant,
each edge (line) represents having less than X mismatches, as defined during the `haplogroup()`
step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$plotGrouping(main = NULL, vertex.size = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{main}}{title for the plot}

\item{\code{vertex.size}}{size of the vertices (dots). Defaults to 3. Useful if you want
to better see the grouping relationships}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, but prints a plot. To save the plot use jpeg(), png() or similar functions.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$haplogroup(error = 0.02)
myBlock$plotGrouping("My plot")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-alleles"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-alleles}{}}}
\subsection{Method \code{alleles()}}{
Function to obtain allele codes for each individual
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$alleles(type = "allele")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{one of the following: "allele", "variant", "haplotype", "sequence".
Admits partial matching (e.g. "hap" instead of "haplotype"). See details for more information.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Each type produces a different encoding of alleles:
- allele: binary code where 0 means reference allele and 1 means alternative allele.
For example: 00010, 00110, 11110 ...
- variant: alphanumeric variant codes, where each unique variant is given a
specific name. For example: V010, V011, V123...
- haplotype: alphanumeric haplotype codes, based on a pre-calculated haplotype grouping
using `haplogroup()`. For example: H01, H02, H03... Haplotype numbers are based
on frequency (H01 is always the most frequent).
- sequence: DNA sequences based on reference and alternative alleles, as well as
on a reference sequence. Can only be used after using the method `buildSequences()`.
For example: "AACCTTTG", "AACCATTG", "AACCATTC"...
}

\subsection{Returns}{
A named vector containing allele codes for each chromosome of each
individual. Each individual name is repeated `ploidy` times (e.g. 2 times for a diploid)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
myBlock$alleles("var")

myBlock$haplogroup(error = 0.02)
myBlock$alleles("hap")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-buildSequences"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-buildSequences}{}}}
\subsection{Method \code{buildSequences()}}{
Method to turn phased haplotypes into sequences
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$buildSequences(sequence, start, pos = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequence}}{character vector containing a reference DNA string.}

\item{\code{start}}{coordinates (in the reference genome) where the string begins. This
is related to the 'pos' field which is often related to SNP markers.}

\item{\code{pos}}{data.frame, required if not provided during the initialization of
the `hapBlock` object. Must contain columns "chrom", "pos", "ref", "alt" insensitive
to capitalization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Does not return anything, but `myBlock$alleles("sequence")` will now
be available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-mismatch"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-mismatch}{}}}
\subsection{Method \code{mismatch()}}{
Returns expected number of mismatches between two identical haplotypes (in this haploBlock)
given a specific genotyping error rate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$mismatch(error)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error}}{numeric, error rate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector with three values, lc (lower confidence interval), exp (average
number of expeted mismatches), uc (upper confidence interval).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myBlock <- hapBlock$new(exPhase[1:10,], exInds, ploidy = 2)
expected <- myBlock$mismatch(error = c(0.01))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hapBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hapBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hapBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
