% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{numerate}
\alias{numerate}
\title{Utility function to produce alphanumeric codes}
\usage{
numerate(x, prefix = "V")
}
\arguments{
\item{x}{vector of objects to generate labels for}

\item{prefix}{character to be used as prefix for the labels}
}
\description{
It takes any set of numbers and returns a 0-padded
alphanumeric code with a specificed prefix
}
